+++ callback/fabrick/annuity_single_rate.go
@@ -1,14 +1,12 @@
-package payment
+package fabrick
 
 import (
 	"encoding/json"
 	"log"
 	"net/http"
 	"strings"
-	"time"
 
 	"github.com/wopta/goworkspace/lib"
-	"github.com/wopta/goworkspace/mail"
 	"github.com/wopta/goworkspace/models"
 	"github.com/wopta/goworkspace/network"
 	plc "github.com/wopta/goworkspace/policy"
@@ -16,11 +14,10 @@ import (
 
 /*
 This handler is intended to handle all callbacks from fabrick that represent
-a transaction that is the first in it's annuity. Other then pay the related
-transaction, it should also update the policy state, and promote certain data
-if it is the first transaction ever
+a transaction that is in an already valid policy annuity. It should only pay the
+transaction and have no other side effects
 */
-func AnnuityFirstRateFx(_ http.ResponseWriter, r *http.Request) (string, any, error) {
+func AnnuitySingleRateFx(_ http.ResponseWriter, r *http.Request) (string, interface{}, error) {
 	var (
 		err            error
 		requestPayload FabrickRequestPayload
@@ -28,7 +25,7 @@ func AnnuityFirstRateFx(_ http.ResponseWriter, r *http.Request) (string, any, er
 		response       = FabrickResponse{Result: true, Locale: "it"}
 	)
 
-	log.SetPrefix("[AnnuityFirstRateFx] ")
+	log.SetPrefix("[AnnuitySingleRateFx] ")
 	defer log.SetPrefix("")
 
 	log.Println("Handler start -----------------------------------------------")
@@ -65,7 +62,7 @@ func AnnuityFirstRateFx(_ http.ResponseWriter, r *http.Request) (string, any, er
 
 	paymentMethod = strings.ToLower(request.Bill.Transactions[0].PaymentMethod)
 
-	err = annuityFirstRate(policyUid, providerId, trSchedule, paymentMethod)
+	err = annuitySingleRate(policyUid, providerId, trSchedule, paymentMethod)
 	if err != nil {
 		log.Printf("error paying first annuity rate: %s", err)
 		response.Result = false
@@ -79,12 +76,11 @@ func AnnuityFirstRateFx(_ http.ResponseWriter, r *http.Request) (string, any, er
 	return string(stringRes), response, nil
 }
 
-func annuityFirstRate(policyUid, providerId, trSchedule, paymentMethod string) error {
+func annuitySingleRate(policyUid, providerId, trSchedule, paymentMethod string) error {
 	var (
 		policy      models.Policy
 		transaction models.Transaction
 		networkNode *models.NetworkNode
-		warrant     *models.Warrant
 		err         error
 	)
 
@@ -94,55 +90,12 @@ func annuityFirstRate(policyUid, providerId, trSchedule, paymentMethod string) e
 	}
 
 	networkNode = network.GetNetworkNodeByUid(policy.ProducerUid)
-	if networkNode != nil {
-		warrant = networkNode.GetWarrant()
-	}
 
 	if transaction, err = payTransaction(policyUid, providerId, trSchedule, paymentMethod, networkNode); err != nil {
 		return err
 	}
 
-	policy.IsPay = true
-	policy.Status = models.PolicyStatusPay
-	policy.StatusHistory = append(policy.StatusHistory, models.PolicyStatusPay)
-	policy.Updated = time.Now().UTC()
-
-	if policy.Annuity == 0 {
-		// TODO: all methods save the data, they shouldn't to avoid data corruption
-		if err = plc.SetUserIntoPolicyContractor(&policy, ""); err != nil {
-			return err
-		}
-
-		if err = plc.AddContract(&policy, ""); err != nil {
-			return err
-		}
-
-		if err = network.UpdateNetworkNodePortfolio("", &policy, networkNode); err != nil {
-			return err
-		}
-
-		flowName, _ := policy.GetFlow(networkNode, warrant)
-		toAddress := mail.Address{}
-		ccAddress := mail.Address{}
-		fromAddress := mail.AddressAnna
-
-		switch flowName {
-		case models.ProviderMgaFlow:
-			toAddress = mail.GetContractorEmail(&policy)
-			ccAddress = mail.GetNetworkNodeEmail(networkNode)
-		case models.RemittanceMgaFlow:
-			toAddress = mail.GetNetworkNodeEmail(networkNode)
-		case models.MgaFlow, models.ECommerceFlow:
-			toAddress = mail.GetContractorEmail(&policy)
-		}
-
-		mail.SendMailContract(policy, nil, fromAddress, toAddress, ccAddress, flowName)
-	}
-
 	firestoreBatch := map[string]map[string]interface{}{
-		lib.PolicyCollection: {
-			policy.Uid: policy,
-		},
 		lib.TransactionsCollection: {
 			transaction.Uid: transaction,
 		},
@@ -150,7 +103,6 @@ func annuityFirstRate(policyUid, providerId, trSchedule, paymentMethod string) e
 	if err = lib.SetBatchFirestoreErr(firestoreBatch); err != nil {
 		return err
 	}
-	policy.BigquerySave("")
 	transaction.BigQuerySave("")
 
 	return nil
